OPEN - issue DNA-14: SequencerSystem does not use Workspace and Session in a threadsafe manner http://jira.jboss.org/jira/browse/DNA-14
Refactored the Sequencing framework to use a new ISessionFactory interface to look up workspaces by name and to create new JCR Sessions when needed.  The WorkspaceListeners each use a separate session, and the sequencing of a node is done by creating a session, running the sequencers for the node in the session, saving the session, and then logging out of the session.

All of the items in the queues are now immutable, and other classes (like WorkspaceListener) are now thread safe.  The ability to safely monitor and sequence should now be greatly improved.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@41 76366958-4244-0410-ad5e-bbfabb93f86b