Refactored the general capability to have a library of components into dna-common, and changed the SequencerLibrary to reuse that.  Analyses will also use this, as will other components that rely upon loading Java classes that implement some interface.
Also extracted the monitoring of JCR workspaces into a separate component to which EventListeners and NodeChangeListeners can be added and which centralizes the processing and dissemination of JCR events.  SequencerService was changed to be a NodeChangeListener, and other components (like that which monitors the JCR repository to maintain the various ComponentLibrary instances) will also make use of it.

Finally, moved some of the dna-service components around (removed 'o.j.dna.services.util' package, moving items into the 'o.j.dna.services' package).

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@63 76366958-4244-0410-ad5e-bbfabb93f86b