DNA-21: Support internationalization and localization http://jira.jboss.com/jira/browse/DNA-21
Additional logging and internationalization refinements.  It is not possible to set the locale for the logging if the VM's default locale is not to be used.  Also, the I18n class now has methods that take a specific locale into which a message should be localized (using the VM's default locale if none is provided).

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@76 76366958-4244-0410-ad5e-bbfabb93f86b