DNA-29: Create abstract base class for sequencers that process streamed contents and generate simple data http://jira.jboss.com/jira/browse/DNA-29
Extracted into the StreamSequencer abstract base class the common logic for sequencng:  get the stream, delegate to an abstract method to process the stream and populate a Map<Path,Object> of properties and values, close the stream, find/create all output nodes, write the output property map to each output node, and save the repository session.

This required a fairly complete and useful "Path" class, which was placed in common.  Several other classes were modified to support testing.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@86 76366958-4244-0410-ad5e-bbfabb93f86b