DNA-29: Create abstract base class for sequencers that process streamed contents and generate simple data http://jira.jboss.com/jira/browse/DNA-29
Refactored how the StreamSequencer subclasses write their output.  The new approach uses an interface, and the implementation organizes the properties by node, simplifying the logic required to process the properties and save them to JCR.  Using the interface also makes it possible (in the future) to implement the output to write the data immediately.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@91 76366958-4244-0410-ad5e-bbfabb93f86b