DNA-60: Progress monitor framework does not allow for localization of progress statuses http://jira.jboss.org/jira/browse/DNA-60
Changed the "getStatus()" method in ProgresMonitor to be "getStatus(Locale)" that can take a null locale when the default locale should be used.  All implementation classes were affected (this is before 0.1 release, so that's acceptable).  Corrected how the SimpleProgressMonitor handles the taskName and taskNameParameters.  Also Added test cases for when the monitor has not yet been started as well as after it has been started (and given a task name).  The latter is done by supplying different locales (with and without parameters).

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@109 76366958-4244-0410-ad5e-bbfabb93f86b