DNA-77: SequencerOutput setProperty and setReference method signatures need to be changed http://jira.jboss.org/jira/browse/DNA-77
Added two methods to the SequencerOutput interface:  a method to obtain the org.jboss.dna.spi.graph.ValueFactories instance that can be used to create paths, names, and other value objects (like binary objects, date-time instants, strings, doubles, etc.); and a method to set the property values using a Path object for the node path, and a Name object for the property name.  The Path, Name, DateTime, Reference, Binary, ValueFactory, and ValueFactories interfaces are all part of the SPI, and will be used by the federation engine, connectors, and (now) sequencers.  The existing SequencerOutput methods (that take string paths and names) simply create the Path and Name objects and then delegate to the new method.


git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@223 76366958-4244-0410-ad5e-bbfabb93f86b