DNA-82: Break TextEncoder interface into two interfaces: one for encoding and one for decoding http://jira.jboss.org/jira/browse/DNA-82
Broke into separate interfaces, and changed all uses.  Some places did require references to both a TextEncoder and TextDecoder, but most only required a single reference.  It also made some of the uses of a decoder more clear (than calling it a "decoder" but having the type be TextEncoder, as it was previously.)

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@235 76366958-4244-0410-ad5e-bbfabb93f86b