DNA-135: Time-based stopwatch tests sometimes fail on slow machines http://jira.jboss.org/jira/browse/DNA-135
I suspect the most likely cause is the call to "Thread.sleep".  These total number of milliseconds was pretty small and probably too close to the precision of the system clock, so I increased the value by an order of magnitude.  Hopefully this will improve the likelihood of the tests passing.  If not, we'll need to test these using a different technique, or ignore these tests.  Greg, please comment with your results (and please reopen if the change is not successful).

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@292 76366958-4244-0410-ad5e-bbfabb93f86b