DNA-115 - Create federation service http://jira.jboss.com/jira/browse/DNA-115
Added more tests for FederationService, including tests for those methods that read the configuration repository.  To allow the configuration of a FederatedRepository to be updated in a concurrent environment, the configuration-related information (regions, default cache policy) was pulled into a separate FederatedRepositoryConfig object.  A FederatedRepository's configuration can therefore be replaced wholesale.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@324 76366958-4244-0410-ad5e-bbfabb93f86b