DNA-173 - Create reusable PathExpression class in the SPI http://jira.jboss.com/jira/browse/DNA-173
Moved the path expression functionality out of the SequencerPathExpression (which now contains a PathExpression for the input along with an output expression) and into PathExpression in the SPI.  Added several test cases, including some new cases.  There is no change in the behavior, and all existing test cases still work.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@349 76366958-4244-0410-ad5e-bbfabb93f86b