DNA-163 - Most, if not all, of the assertions in BasicMultiValueProperty are incorrect. http://jira.jboss.com/jira/browse/DNA-163
The assertions were indeed correct, but they were misplaced and should instead be in the constructors.  The BasicEmptyProperty class should be used for properties that have no values; the BasicSingleValueProperty should be used for properties that have just a single value (which may be null); and the BasicMultiValueProperty should be used for properties that contain 2+ values.  This is done to minimize the memory footprint: if a single Property implementation were used, it would have to support multiple values, but in reality most properties will be properties with one a single value and therefore the extra collection is simply wasted memory.  By having different implementations for the different situations, we're able to hold onto the values with the most efficiency.  Also, since the objects are intended to be immutable, there is no need to support changing the values.  This description was added to the JavaDoc.

This relies upon a new ArgCheck method, which was added with tests (along with several other similar methods).

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@351 76366958-4244-0410-ad5e-bbfabb93f86b