DNA-189 - Change RepositorySource to extend the ManagedRepositoryConnectionFactory interface, so it has management methods http://jira.jboss.com/jira/browse/DNA-189
The RepositorySource only has methods for getting/setting properties and creating connections. This means that managing the set of connections is not part of the interface, and this results in the RepositorySource instances needing to be wrapped by RepositoryConnectionPools. This makes it more difficult (and less consistent) to manage the instances.

Change name of RepositoryConnectionPool interface to ManagedRepositoryConnectionFactory
Change RepositorySource to extend the ManagedRepositoryConnectionFactory interface (rather than RepositoryConnectionFactory)
Change name of BasicRepositoryConnectionPool to RepositoryConnectionPool (since there is no "pool" interface anymore)
Create an AbstractRepositorySource implementation that uses a pool to manage the instances
Change the existing RepositorySource implementations to extend AbstractRepositorySource (this is not required, but it does encapsulate the management of the connections)
Clean up RepositorySourceManager, since it no longer needs to distinguish between RepositorySource and pools (since all RepositorySource are now manageable)

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@370 76366958-4244-0410-ad5e-bbfabb93f86b