DNA-115 - Create federation service http://jira.jboss.com/jira/browse/DNA-115
Renamed the "getEnvironment()" methods to "getExecutionContext()" to reflect the previous change in name of the ExecutionContext (was ExecutionEnvironment).  Also renamed the "getProperties()" method of some of the SPI graph commands that returned an Iterator<Property> to now be "getPropertyIterator()", which is now consistent with the other graph commands.  Made FederatedNode a concrete class (not an interface), moved it, and implemented additional interfaces so that it can be treated as a GetNodeCommand.  Finally, made more changes to the FederationCommandExecutor implementation.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@379 76366958-4244-0410-ad5e-bbfabb93f86b