DNA-188 - Complete the federating command executor http://jira.jboss.com/jira/browse/DNA-188
Needed to add the ability to get the identification properties (e.g., a UUID property) for each child.  These properties are required when getting nodes from the original sources so that the merge logic can determine whether child nodes (from different sources) represent the same node or different nodes.  These properties are also required in the Federation Connector to obtain the UUID property for each node in the integrated repository cache.  The GetChildrenCommand was modified to remove most of the "setChildren(...)" methods and to instead have "addChild(Segment, Property)" methods.  The existing connectors were modified to use this new capability.

To support storing UUIDs as properties, a UUID property type was added and a new UuidFactory (which extends ValueFactory<UUID>) interface was defined.  Each existing factory interface (and implementation) was then modified to add a "create(UUID value)" method.  All ValueFactories implementations were modified as well.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@380 76366958-4244-0410-ad5e-bbfabb93f86b