DNA-196 - Add methods to ValueFactory that takes Iterator<?> and returns Iterator<T> http://jira.jboss.com/jira/browse/DNA-196
Added a single "create(Iterator<?> values):Iterator<T>" to the ValueFactory interface, and added implementation methods to all concrete classes.  The AbstractValueFactory class defines a ConvertingIterator<T> that wraps an Iterator<?> and simply calls "create(Object):T" on the value factory that is using it.  Thus, this method is actually quite efficient, since the only cost is the wrapped iterator instance.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@400 76366958-4244-0410-ad5e-bbfabb93f86b