DNA-192 - Clean up the repository connector SPI http://jira.jboss.com/jira/browse/DNA-192
Refactored the classes again, since the RepositorySource needs to be serializable and able to be turned into a JNDI reference. In other words, the RepositorySource needs to be a factory for creating connections to a specific repository, but it shouldn't manage the pool of connections (since the pool and any live connections cannot be serialized).

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@403 76366958-4244-0410-ad5e-bbfabb93f86b