DNA-83 - Federate content from JBoss Cache instance(s) http://jira.jboss.com/jira/browse/DNA-83
Addressed the need to support same-name siblings, which is more difficult to do with JBoss Cache, since children are not kept in any order.  The code now maintains for any node in JBoss Cache a Path.Segment[] in a property, which stores it's ordered list of children.  This is the most memory-efficient way to do this (that I could think of), since the actual Path.Segment objects should already be stored in the Fqn for each node, and no indexes need be maintained.  So the only overhead is that associated with the array (roughly one reference per child).  (Another option considered were wrapping the Path.Segment implementation with another implementation of Path.Segment that also stored the index, but this required a reference plus an integer for each child.)

Found a bug in the AbstractCommandExecutor's handling of GetNodeCommand, which resulted in the GetPropertiesCommand and GetChildrenCommand being executed twice.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@414 76366958-4244-0410-ad5e-bbfabb93f86b