DNA-203 - Unit tests for dna-example-sequencers don't work http://jira.jboss.com/jira/browse/DNA-203
This must have regressed when we upgraded to Jackrabbit 1.4.5, but the unit tests for the examples were apparently not included in the continuous builds.  (See DNA-204 for that.)  We're blindly setting the "jcr:primaryType" property (along with all other properties) on all the nodes, but since that's supplied when creating the node, resetting the "jcr:primaryType" property is not necessary (and apparently no longer allowed).  Added a check in the StreamSequencerAdapter to skip (re)setting this property on each node.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@436 76366958-4244-0410-ad5e-bbfabb93f86b