DNA-204 Unit tests for dna-example-sequencers are not being executed in the Maven builds in the nightly Hudson jobs http://jira.jboss.com/jira/browse/DNA-204
Added the example project(s) to the default profile for the Maven builds. We're getting to the point where the unit tests take a while to run, so we may want to create a reduced set of the projects for the developers to run locally.  (I don't think we're there yet.)

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@437 76366958-4244-0410-ad5e-bbfabb93f86b