DNA-188 - Complete the federating command executor http://jira.jboss.com/jira/browse/DNA-188
Renamed the StandardMergeStrategy to SimpleMergeStrategy, and fully implemented and tested it to merge multiple contributions.  This merge strategy may not be sufficient for many uses, but it works now and does merge properties.  Note that while it does not merge any children, it does ensure that any same-name siblings (from all contributions) have the proper indexes.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@470 76366958-4244-0410-ad5e-bbfabb93f86b