DNA-222 - Create Getting Started example application for repositories http://jira.jboss.com/jira/browse/DNA-222
Created a simple application (subproject under getting started examples) that allows a user to pick from a list of repositories so they can interactively navigate and explore them.  Currently, only a console user interface is implemented.  Several repositories are used: a configuration repository (which defines the others), two in-memory repositories ("Cars" and "Airplanes"), a federated repository ("Vehicles" that federates "Cars" and "Airplanes" below a "vehicles" root), and a cache repository used by "Vehicles".

Most of the unit testing is complete, and several issues have been fixed in other components.  However, there still are a few outstanding issues as well as testing of the interactive mode.  So, still a bit more work to do.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@515 76366958-4244-0410-ad5e-bbfabb93f86b