DNA-224 - dna-connector-federation: one test in FederatingCommandExecutorTest doesn't work with JDK 1.6.0_05 on Mac OS X http://jira.jboss.com/jira/browse/DNA-224
The problem appeared because of a modified implementation of a Set, which resulted in a different order to some of the top-level nodes in a Projection.  This in fact was a bug where the top-level nodes of a projection were not ordered.  This was fixed by making them ordered, keeping the same order as the projection's rules.

All tests now pass on JDK 1.5 and JDK 1.6.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@520 76366958-4244-0410-ad5e-bbfabb93f86b