DNA-222 - Create Getting Started example application for repositories http://jira.jboss.com/jira/browse/DNA-222
Changed how the RepositorySource instances are initialized, removing the need to use JNDI for the FederatedRepositorySource (since the ExecutionContextFactory and RepositoryConnectionFactory are not serializable and could not be placed inside JNDI); instead, RepositoryLibrary (formerly RepositorySourceManager) now initializes each RepositorySource with a RepositoryContext when the RepositorySource is added to the library.  This completely removes the need to use JNDI to look up DNA components, since they are all injected into the correct locations as needed.

Completed more unit testing.  Still lacking tests that use JCR implementation, but that will follow shortly.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@521 76366958-4244-0410-ad5e-bbfabb93f86b