DNA-213 - Review the graph SPI commands to add missing commands and better handle node identification (e.g., when getting children) http://jira.jboss.com/jira/browse/DNA-213
Finished the conversion from the old GraphCommand classes to the newer Request classes.  This involved a very large number of changes, some of which changed from using commands to using the new Graph API.  The new Request classes do a much better job of recording identification properties of the nodes added as results to the requests.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@572 76366958-4244-0410-ad5e-bbfabb93f86b