DNA-242 - Change the XML importer code to no longer use XmlSequencer http://jira.jboss.com/jira/browse/DNA-242
Added new XmlHandler implementation (which subclasses DefaultHandler2 and overrides the appropriate methods) with a unit test that verifies the behavior.  It is designed to be used with a regular org.xml.sax.XMLReader implementation, but processes the XML content and creates the nodes that represent that content.  And while the current component should be perfect for importing XML content into a repository, it can also be subclassed with other DefaultHandler2 methods being overridden to handle other parts of the XML content, including XML comments, DTD entities, entity resolution, etc.  This approach of specializing a DefaultHandler2 implementation is much more extensible and customizable than providing an "importer" component.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@590 76366958-4244-0410-ad5e-bbfabb93f86b