DNA-245 - Add methods to Name, Path.Segment, and Path to allow using custom delimiters http://jira.jboss.com/jira/browse/DNA-245
Added to Name, Path.Segment, and Path the "getString(NamespaceRegistry, TextEncoder, TextEncoder)" method, which uses the 2nd encoder to preprocess the various delimiters.  These methods can be used, for example, to escape (or convert) the delimiters in a Path (for example, to escape the ":" as needed by a file system, or to convert "/" to a different delimiter.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@595 76366958-4244-0410-ad5e-bbfabb93f86b