Removed (unused) methods to determine if error, info, or warning messages are enabled.  The code using the logger should never be concerned with this, as it should just log the messages at the appropriate level.
There still are methods to determine if debug and trace messages are enabled.  However, since these methods are intended for developers (note that they do not take I18n but take strings) and because such methods may require doing more work to prepare the messages (e.g., processing and accumulating information that you'd normally not do), it is still useful to know whether debug or trace is enabled.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@596 76366958-4244-0410-ad5e-bbfabb93f86b