DNA-242 - Change the XML importer code to no longer use XmlSequencer http://jira.jboss.com/jira/browse/DNA-242
Enhanced the XmlHandler to provide a local NamespaceRegistry that mirrors the XML namespaces in the document, and various factories in the XmlHandler that use this local namespace registry.  The benefit is that XmlHandler (and subclasses) can simply just use the factories to process any names that use the XML namespace prefixes.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@603 76366958-4244-0410-ad5e-bbfabb93f86b