DNA-248 - Simplify the ActivityMonitor framework http://jira.jboss.com/jira/browse/DNA-248
Despite building the simplified Task framework (see attached patch file), and after several days of thinking about it and a long discussion with John, it seems like this ProgressMonitor/ActivityMonitor/Task framework is just overkill for what we need.  And it about 1900 LOC overkill.  (Really, it started small and grew as we tried to generalize.  Mostly it was me that was driving it.)

Turns out that only the sequencing used this framework, and its arguable whether there's any benefit to monitoring the progress and/or tasks for a singular sequencing activity.  The connector framework doesn't really need the overhead either, especially when individual requests are so tiny and so many.  So there really isn't a lot of similarity between the two data points (sequencers and connectors).  Rather than try to force something to fit (really the purpose of this JIRA issue), it seems prudent to recognize that we don't really have a well-defined need with well-understood requirements, and that the best thing to do is to simply remove the ProgressMonitor/ActivityMonitor framework.

In doing so, the sequencers became a little simpler.  The XML sequencer became a fair amount simpler.  The code base became quite a bit smaller (>2400 LOC smaller).  And we really didn't lose any functionality.  (Sure, we kind of lost cancellation of sequencers, but that's probably not really true, either - it's quite possible that the only way to REALLY cancel a sequencer that's been stuck reading from a stream is to interrupt the thread, and the cancellation status wouldn't do that anyway.)

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@617 76366958-4244-0410-ad5e-bbfabb93f86b