DNA-217 - Connector SPI should have protocol for cancelling operations http://jira.jboss.com/jira/browse/DNA-217
Added to Request a method isCancelled():boolean to check whether the request was cancelled, and another method "cancel():void" to mark a request as cancelled.  Also added to the CompositeRequest the ability to mark the composite (or any contained request) as cancelled, but to have all other requests (including the composite) be also cancelled.  This is done through a single AtomicBoolean instance shared amongst the composite and contained requests.

Also added tests to verify the desired behavior.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@618 76366958-4244-0410-ad5e-bbfabb93f86b