DNA-223 - XML sequencer does not properly handle namespaces of unqualified attributes http://jira.jboss.com/jira/browse/DNA-223
It is not really possible to correctly handle unqualified attributes in a general manner if we consider the XML Schema specification.  This is because an XML schema by default will require that unqualified attributes inherit the element's namespace, and because within any single XML schema it is possible to define one element that requires unqualified attributes inherit the element's namespace while other element definitions require that unqualified attributes use the default namespace.  In other words, a general-purpose sequencer cannot always do the "correct" thing without thoroughly understanding the XML Schema(s) used by an XML document.

Therefore, this has been addressed by changing the XML sequencer to be configurable as to the desired behavior.  Note that this behavior is fixed for a particular document, and still won't properly handle an XML document that uses a schema with a mixture of qualified and unqualified attribute forms.  The default is to use the default namespace (at that position in the document), which is really the default for XML documents.  To make it easy to specify a sequencer by classname that inherits the element's namespace for unqualified attributes, there is also a new InheritingXmlSequencer, which extends XmlSequencer and simply overrides the default setting.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@619 76366958-4244-0410-ad5e-bbfabb93f86b