DNA-231 - org.jboss.dna.graph.xml.XmlSequencerTest failed with jdk 1.6 (not with jdk 1.5) http://jira.jboss.com/jira/browse/DNA-231
Apparently the SAX parser in Java 1.6.0_07 (build 1.6.0_07-b06-153) on OS-X 10.5.5 does not call the handler methods in the proper order for entity references.  For example, in 1.5 (and in 1.6 on Fedora), this will result in calls to 'startEntity(...)', 'characters(...)', and 'endEntity(...)'. Java 1.6.0_07 on OS-X results in 'startEntity(...)', 'endEntity(...)', and , 'characters(...)' where the 'characters' method is called with character content representing the replaced entity reference PLUS the next characters that would normally be supplied in the subsequent call to 'characters'.

This was addressed by changing how the flag used by these methods to record which entity reference is being processed, and by recording the replacement value when the entity declaration is processed (so that we can separate the entity replacement value from the 'extra' characters).  

New test cases were added to properly verify this behavior, and to ensure that the changes work correctly.  Tests were run in both Java 1.5 and 1.6 on both OS-X and on Fedora.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@620 76366958-4244-0410-ad5e-bbfabb93f86b