DNA-40 Persistant storage for information not stored in other repository sources
Started work on this new connector, which uses Java Persistence Architecture (JPA) to work with EJB3 entities stored in a relational database, using Hibernate for the JPA implementation. (Note that some Hibernate extensions are also used.)

The connector is capable of using one of several models (or schemas). The initial model is called Basic, but the connector supports having multiple models (which can be exist in separate packages) and allowing the user to choose via a JavaBean property in the JpaSource object. (Note that the kind of model used in a particular database instance is recorded in the database, and may not be changed. In this case, the JpaSource's model selection is changed to reflect what is in the database. Changing a database's model would require migration, which is not currently supported, but could be provided by a copy utility that operates on a graph by generically copying the content from one source to another.)

Each model is also responsible for creating the RequestProcessor implementation, so the JpaSource and JpaConnection implementations can work with any model.

The Basic model stores for each node a single record containing the node ID (which is a UUID string) and a blob for the serialized properties. Large The Basic model stores for each node a single record containing the node ID (which is a UUID string) and a blob for the serialized properties. Large The Basic model stores for each node a single record containing the node ID (which is a UUID string) and a blob for the serialized properties. Large The Basic model stores for each node a single record containing the node ID (which is a UUID string) and a blob for the serialized properties. Large The Basic model stores for each node a single record containing the node ID (which is a UUID string) and a blob for the serialized properties. Large The Basic model stores for each node a single record containing the node ID (which is a UUID string) andxistsThe Basic model stores for each node a single record containing the node ID (which is a UUID string) and a blob for the serialized properties. Large Th current implementation is probably close to 75% complete. The Basic entities are mostly completed (minor changes are anticipated), and all of the source infrastructure has been completed. Most of the outstanding work is in the BasicRequestProcessor implementation.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@641 76366958-4244-0410-ad5e-bbfabb93f86b