DNA-254 CreateNodeRequest does not allow passing in location of existing parent
I've changed the CreateNodeRequest to take the Location of a parent plus the Name of a child.  All the connectors have been fixed, and all unit tests pass (and I've verified they're all still working correctly).  Also, the implementation of the Graph has been changed to use the new form, but rather than change the interface (to allow passing in the Location of the parent), I've added new methods.  This means that existing usages of the Graph API will not break and will work as before. 

NOTE that this DOES change/correct the API, and that connectors built against the 0.3 release will no longer compile and will need to be changed.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@643 76366958-4244-0410-ad5e-bbfabb93f86b