DNA-40 Persistant storage for information not stored in other repository sources
Implemented most of the connector behavior (just CopyBranchRequest remains, with a number of minor tasks).  Delete now uses an efficient mechanism for finding subgraphs.  The searches are stores in the database (in a "temporary" area), but this makes computing the subgraph very efficient (one self-join on the children's table per LEVEL of subgraph, not per node) and helps to keep most of the information out of memory.  The same mechanism is used for ReadBranchRequest (and will be used for CopyBranchRequest).

A number of methods also use Hibernate bulk HQL statements, which should be faster.  (The subgraph query is just one way in which they're used.)

Note that the "mark and sweep" approach to deleting nodes was abandoned, because the JCR spec requires that repositories generate delete events for every node that is deleted, not just for the top node in the subgraph that was deleted.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@652 76366958-4244-0410-ad5e-bbfabb93f86b