DNA-250  - Creating a path from a parent path and a segment is not fast
Added an optimized check during the isAtOrBelow and isAtOrAbove methods, and removed the caching of the string form of the path (which was cached based upon the first call to any of the getString methods, regardless of the NamespaceRegistry instance used, meaning the cached string might have used no namespace registry and was returned from a getString(...) that took a namespace registry.  Also, the extra overhead of the cached string was not worth the memory.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@678 76366958-4244-0410-ad5e-bbfabb93f86b