DNA-250  - Creating a path from a parent path and a segment is not fast
Minor tweaks to NameValueFactory.create(String,...) to more efficiently determine the pattern (either internal form with full namespace URIs, or traditional prefixed form) that is used to first parse the string.  Before this change, the internal pattern was almost always used first, despite the fact that most names are created with strings containing the prefixed form.

Also changed the BasicName constructor so that strings representing URIs are interned, since there shouldn't be too many of those, and we really want there to be one string instance for each namespace shared by all the names.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@682 76366958-4244-0410-ad5e-bbfabb93f86b