DNA-266 Improve usability of graph API with additional methods
Added an easy way to set references to nodes as property values.  Any Node or Location passed as a property value will automatically be converted to a Reference property value pointing to the UUID of the Node (or Location).  Also added additional "to" methods (for setting property values) that take a Node and Location, making it more obvious what values are allowed.

Also changed the SetValuesTo interface so that it's possible to specify the "on" before the values.  For example:
        graph.set("propName").to("value").on("/node1");
        graph.set("propName").on("/node1").to("value");
The first example was what was original required, but the bottom one also works now.  Notice that the 3rd method called in each of these ("on" in the first, "to" in the second example) returns an interface that completes the set operation.  In other words, the following results in a compile error:
        graph.set("propName").to("value").on("/node1").to("value2");
This is great, since it is the desired behavior.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@692 76366958-4244-0410-ad5e-bbfabb93f86b