DNA-40
Per JSR-283 section 5.1.7, corrected this connector's copy behavior for how "internal" references (from nodes within the subgraph to nodes also within the subgraph) are remapped in the copy, so that the copy's references are also internal to the copy.  Also verified that "outward" references (from nodes within the subgraph to nodes outside of the subgraph) are copied correctly, and that "inward" references (from nodes outside of the original subgraph to nodes in the original subgraph) are not affected by the copy operation (since the original subgraph is not modified).

Note that JSR-170 does not define the semantics of how references are handled in a copy operation.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@693 76366958-4244-0410-ad5e-bbfabb93f86b