DNA-266 Improve usability of graph API with additional methods
Added additional methods for setting the values of properties, overloading with the different types allowed for property values (e.g., boolean, int, Path, Name, etc.).  This resulted in quite a bit more code (that is all pretty straightforward), but this detriment is outweighed by the API being more useful and clear.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@694 76366958-4244-0410-ad5e-bbfabb93f86b