DNA-260 Add factory method to create a Binary instance from a (local) File
Created the BinaryFactory interface, which extends ValueFactory<Binary> and adds several methods for creating Binary objects from File objects, and from InputStream (or Reader) when then binary's hash is already known.  These latter methods will allow an implementation to maintain a pool of Binary objects keyed by their hash, so that only one Binary object per unique value is needed.  Also added a method to find a Binary value given the hash (so that connectors and clients that know the hash may attempt to reuse an existing Binary value with the same hash, without having to create the InputStream (or Reader), since that may be expensive even if the stream/reader is not used.

None of the factories in 'org.jboss.dna.graph' reuse Binary values, however; they always create a new object.

Also added some better error handling logic in IoUtil - getting an exception when closing a stream (in the finally block) will not stomp on an exception thrown from the try block.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@698 76366958-4244-0410-ad5e-bbfabb93f86b