DNA-264 Connectors may need to determine the MIME type for content
This change adds the ability in ExecutionContext to get a MimeTypeDetector, which can be used to get the MIME type for some content.  MimeTypeDetectors (the library component in 'org.jboss.dna.repository') now implements this interface (the signature already matched), so that can be used in an ExecutionContext implementation in the complete DNA system.

Also added a new ExtensionBasedMimeTypeDetector implementation that reads from the classloader a properties file defining the extensions for each MIME type.  A file containing many common extensions is provided, although a user can either supply their own or use a different constructor that allows them to pass the mappings in as a parameter.

The BasicExecutionContext uses this ExtensionBasedMimeTypeDetector as the default implementation, although its possible to use a different one.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@700 76366958-4244-0410-ad5e-bbfabb93f86b